requirements.txt
fastapi
pillow
joblib
tensorflow==2.11.0
opencv-python
numpy
splitfolders
pandas
tqdm